TL/MHBA: p2p asr sync
