+++ src/components/tl/mhba/tl_mhba.c
@@ -18,6 +18,9 @@ static ucc_config_field_t ucc_tl_mhba_lib_config_table[] = {
     {"TRANSPOSE", "0", "Boolean - with transpose or not",
      ucc_offsetof(ucc_tl_mhba_lib_config_t, transpose), UCC_CONFIG_TYPE_UINT},
 
+    {"ASR_BARRIER", "0", "Boolean - use  service barrier or p2p sync of ASRs",
+     ucc_offsetof(ucc_tl_mhba_lib_config_t, asr_barrier), UCC_CONFIG_TYPE_UINT},
+
     {"TRANPOSE_BUF_SIZE", "128k", "Size of the pre-allocated transpose buffer",
      ucc_offsetof(ucc_tl_mhba_lib_config_t, transpose_buf_size),
      UCC_CONFIG_TYPE_MEMUNITS},
